Fix bug with variable composition
This PR fixes an issue with variable composition where an error would be thrown if the local variable's name did not match the class property's name.

Prior to this fix, the following test would error because foo is does not have the same name as subscription.
```
let foo = of("foo").subscribe(value => this.value = value);
this.subscription.add(foo);
```