Update package.json to point to NPM registry instead of GitHub (#5)
This PR updates the package.json to point to the NPM registry. I
attempted to use the GitHub package registry, but this requires special
`.yarnrc` configuration in the project as well as authenticating against
the registry. Anonymous users are not supported for GitHub Registry.

We now have our own NPM scope: `@smartpassk12`. `@smartpass` was taken
:(

You can publish a new version of this package by following the
instructions in the README.